+++ apps/pokemon-service/src/commands/pokemon/domain/value-objects/base-stats/base-stats-sp-attack.ts
@@ -2,7 +2,7 @@ import { z } from 'zod';
 import { ValueObject } from '../../../../../common/domain/value-object';
 import { InvalidPokemonException } from '../../exceptions/invalid-pokemon.exception';
 
-const BaseStatsSpAttackSchema = z.number().min(0);
+const BaseStatsSpAttackSchema = z.number().int().min(0);
 
 /**
  * 種族値 とくこう
