+++ apps/pokemon-service/src/commands/pokemon/domain/value-objects/base-stats/base-stats-sp-defense.ts
@@ -2,7 +2,7 @@ import { z } from 'zod';
 import { ValueObject } from '../../../../../common/domain/value-object';
 import { InvalidPokemonException } from '../../exceptions/invalid-pokemon.exception';
 
-const BaseStatsSpDefenseSchema = z.number().min(0);
+const BaseStatsSpDefenseSchema = z.number().int().min(0);
 
 /**
  * 種族値 とくぼう
