+++ apps/pokemon-service/src/commands/pokemon/domain/value-objects/base-stats/base-stats-speed.ts
@@ -2,7 +2,7 @@ import { z } from 'zod';
 import { ValueObject } from '../../../../../common/domain/value-object';
 import { InvalidPokemonException } from '../../exceptions/invalid-pokemon.exception';
 
-const BaseStatsSpeedSchema = z.number().min(0);
+const BaseStatsSpeedSchema = z.number().int().min(0);
 
 /**
  * 種族値 すばやさ
